+++ lib/kafka/client.rb
@@ -71,6 +71,21 @@ module Kafka
       @cluster = initialize_cluster
     end
 
+    # Delivers a single message to the Kafka cluster.
+    #
+    # **Note:** Only use this API for low-throughput scenarios. If you want to deliver
+    # many messages at a high rate, or if you want to configure the way messages are
+    # sent, use the {#producer} or {#async_producer} APIs instead.
+    #
+    # @param value [String, nil] the message value.
+    # @param key [String, nil] the message key.
+    # @param topic [String] the topic that the message should be written to.
+    # @param partition [Integer, nil] the partition that the message should be written
+    #   to, or `nil` if either `partition_key` is passed or the partition should be
+    #   chosen at random.
+    # @param partition_key [String] a value used to deterministically choose a
+    #   partition to write to.
+    # @return [nil]
     def deliver_message(value, options={})
       key = options[:key]
       topic = options[:topic]
@@ -214,6 +229,7 @@ module Kafka
         delivery_threshold: delivery_threshold,
         max_queue_size: max_queue_size,
         instrumenter: @instrumenter,
+        logger: @logger,
       )
     end
 
@@ -410,7 +426,7 @@ module Kafka
 
         batches.each do |batch|
           batch.messages.each(&block)
-          offsets[batch.partition] = batch.last_offset
+          offsets[batch.partition] = batch.last_offset + 1
         end
       end
     end