+++ lib/kafka/datadog.rb
@@ -96,6 +96,7 @@ module Kafka
 
         tags = {
           client: event.payload.fetch(:client_id),
+          group_id: event.payload.fetch(:group_id),
           topic: event.payload.fetch(:topic),
           partition: event.payload.fetch(:partition),
         }
@@ -115,6 +116,7 @@ module Kafka
 
         tags = {
           client: event.payload.fetch(:client_id),
+          group_id: event.payload.fetch(:group_id),
           topic: event.payload.fetch(:topic),
           partition: event.payload.fetch(:partition),
         }
@@ -134,6 +136,7 @@ module Kafka
       def produce_message(event)
         client = event.payload.fetch(:client_id)
         topic = event.payload.fetch(:topic)
+        message_size = event.payload.fetch(:message_size)
         buffer_size = event.payload.fetch(:buffer_size)
         max_buffer_size = event.payload.fetch(:max_buffer_size)
         buffer_fill_ratio = buffer_size.to_f / max_buffer_size.to_f
@@ -144,6 +147,8 @@ module Kafka
 
         # This gets us the write rate.
         increment("producer.produce.messages", tags: tags.merge(topic: topic))
+
+        histogram("producer.produce.message_size", message_size, tags: tags.merge(topic: topic))
 
         # This gets us the avg/max buffer size per producer.
         histogram("producer.buffer.size", buffer_size, tags: tags)
@@ -206,6 +211,37 @@ module Kafka
       end
 
       attach_to "producer.kafka"
+    end
+
+    class AsyncProducerSubscriber < StatsdSubscriber
+      def enqueue_message(event)
+        client = event.payload.fetch(:client_id)
+        topic = event.payload.fetch(:topic)
+        queue_size = event.payload.fetch(:queue_size)
+        max_queue_size = event.payload.fetch(:max_queue_size)
+        queue_fill_ratio = queue_size.to_f / max_queue_size.to_f
+
+        tags = {
+          client: client,
+        }
+
+        # This gets us the avg/max queue size per producer.
+        histogram("async_producer.queue.size", queue_size, tags: tags)
+
+        # This gets us the avg/max queue fill ratio per producer.
+        histogram("async_producer.queue.fill_ratio", queue_fill_ratio, tags: tags)
+      end
+
+      def buffer_overflow(event)
+        tags = {
+          client: event.payload.fetch(:client_id),
+          topic: event.payload.fetch(:topic),
+        }
+
+        increment("async_producer.produce.errors", tags: tags)
+      end
+
+      attach_to "async_producer.kafka"
     end
   end
 end