+++ lib/kafka/offset_manager.rb
@@ -17,7 +17,8 @@ module Kafka
       @processed_offsets = {}
       @default_offsets = {}
       @committed_offsets = nil
-      @last_commit = Time.at(0)
+      @resolved_offsets = {}
+      @last_commit = Time.now
     end
 
     def set_default_offset(topic, default_offset)
@@ -27,7 +28,13 @@ module Kafka
     def mark_as_processed(topic, partition, offset)
       @uncommitted_offsets += 1
       @processed_offsets[topic] ||= {}
-      @processed_offsets[topic][partition] = offset + 1
+      @processed_offsets[topic][partition] = offset
+      @logger.debug "Marking #{topic}/#{partition}:#{offset} as committed"
+    end
+
+    def seek_to_default(topic, partition)
+      @processed_offsets[topic] ||= {}
+      @processed_offsets[topic][partition] = -1
     end
 
     def next_offset_for(topic, partition)
@@ -38,20 +45,20 @@ module Kafka
       # A negative offset means that no offset has been committed, so we need to
       # resolve the default offset for the topic.
       if offset < 0
-        offset = @default_offsets.fetch(topic)
-        offset = @cluster.resolve_offset(topic, partition, offset)
-
-        # Make sure we commit this offset so that we don't repeat have to
-        # resolve the default offset every time.
-        mark_as_processed(topic, partition, offset - 1)
+        resolve_offset(topic, partition)
+      else
+        # The next offset is the last offset plus one.
+        offset + 1
       end
-
-      offset
     end
 
     def commit_offsets
       unless @processed_offsets.empty?
-        @logger.info "Committing offsets for #{@uncommitted_offsets} messages"
+        pretty_offsets = @processed_offsets.flat_map {|topic, partitions|
+          partitions.map {|partition, offset| "#{topic}/#{partition}:#{offset}" }
+        }.join(", ")
+
+        @logger.info "Committing offsets: #{pretty_offsets}"
 
         @group.commit_offsets(@processed_offsets)
 
@@ -63,18 +70,43 @@ module Kafka
     end
 
     def commit_offsets_if_necessary
-      if seconds_since_last_commit >= @commit_interval || commit_threshold_reached?
+      if commit_timeout_reached? || commit_threshold_reached?
         commit_offsets
       end
     end
 
     def clear_offsets
-      @uncommitted_offsets = 0
       @processed_offsets.clear
+
+      # Clear the cached commits from the brokers.
+      @committed_offsets = nil
+    end
+
+    def clear_offsets_excluding(excluded)
+      # Clear all offsets that aren't in `excluded`.
+      @processed_offsets.each do |topic, partitions|
+        partitions.keep_if do |partition, _|
+          excluded.fetch(topic, []).include?(partition)
+        end
+      end
+
+      # Clear the cached commits from the brokers.
       @committed_offsets = nil
     end
 
     private
+
+    def resolve_offset(topic, partition)
+      @resolved_offsets[topic] ||= fetch_resolved_offsets(topic)
+      @resolved_offsets[topic].fetch(partition)
+    end
+
+    def fetch_resolved_offsets(topic)
+      default_offset = @default_offsets.fetch(topic)
+      partitions = @group.assigned_partitions.fetch(topic)
+
+      @cluster.resolve_offsets(topic, partitions, default_offset)
+    end
 
     def seconds_since_last_commit
       Time.now - @last_commit
@@ -83,6 +115,10 @@ module Kafka
     def committed_offset_for(topic, partition)
       @committed_offsets ||= @group.fetch_offsets
       @committed_offsets.offset_for(topic, partition)
+    end
+
+    def commit_timeout_reached?
+      @commit_interval != 0 && seconds_since_last_commit >= @commit_interval
     end
 
     def commit_threshold_reached?